testutil: move testutil.CompareUnorderedStringSlice to new package
